Rework calibration to be more accurate. This one's complicated:
Calibration now fetches a message from the queue, but *does not* delete
it. This is solely to get the time the first-available message so we can
tell how far into the retention window we are.

If there is a message, compute the tweet rate based on the amount of
retention left after accounting for how long the tweet has already sat
there.

If there isn't a message, or we can't parse the timestamp, or anything
else weird, just use the entire retention window to compute the tweet
    rate.